refactor memory
