+++ src/services/userMemory.ts
@@ -18,29 +18,29 @@ import {
 
 class UserMemoryService {
   addContextMemory = async (params: AddContextMemoryParams): Promise<AddContextMemoryResult> => {
-    return lambdaClient.userMemories.toolAddContextMemory.mutate(params);
+    return lambdaClient.userMemories.tools.addContextMemory.mutate(params);
   };
 
   addExperienceMemory = async (
     params: AddExperienceMemoryParams,
   ): Promise<AddExperienceMemoryResult> => {
-    return lambdaClient.userMemories.toolAddExperienceMemory.mutate(params);
+    return lambdaClient.userMemories.tools.addExperienceMemory.mutate(params);
   };
 
   addIdentityMemory = async (params: AddIdentityMemoryParams): Promise<AddIdentityMemoryResult> => {
-    return lambdaClient.userMemories.toolAddIdentityMemory.mutate(params);
+    return lambdaClient.userMemories.tools.addIdentityMemory.mutate(params);
   };
 
   addPreferenceMemory = async (
     params: AddPreferenceMemoryParams,
   ): Promise<AddPreferenceMemoryResult> => {
-    return lambdaClient.userMemories.toolAddPreferenceMemory.mutate(params);
+    return lambdaClient.userMemories.tools.addPreferenceMemory.mutate(params);
   };
 
   removeIdentityMemory = async (
     params: RemoveIdentityMemoryParams,
   ): Promise<RemoveIdentityMemoryResult> => {
-    return lambdaClient.userMemories.toolRemoveIdentityMemory.mutate(params);
+    return lambdaClient.userMemories.tools.removeIdentityMemory.mutate(params);
   };
 
   retrieveMemory = async (params: SearchMemoryParams): Promise<SearchMemoryResult> => {
@@ -48,13 +48,13 @@ class UserMemoryService {
   };
 
   searchMemory = async (params: SearchMemoryParams): Promise<SearchMemoryResult> => {
-    return lambdaClient.userMemories.toolSearchMemory.query(params);
+    return lambdaClient.userMemories.tools.searchMemory.query(params);
   };
 
   updateIdentityMemory = async (
     params: UpdateIdentityMemoryParams,
   ): Promise<UpdateIdentityMemoryResult> => {
-    return lambdaClient.userMemories.toolUpdateIdentityMemory.mutate(params);
+    return lambdaClient.userMemories.tools.updateIdentityMemory.mutate(params);
   };
 }
