+++ src/store/userMemory/slices/base/index.ts
@@ -1,0 +1,1 @@
+export { createBaseSlice, type BaseAction } from './action';
