+++ src/store/userMemory/slices/experience/index.ts
@@ -1,0 +1,1 @@
+export { createExperienceSlice, type ExperienceAction } from './action';
