+++ src/store/userMemory/slices/identity/index.ts
@@ -1,0 +1,1 @@
+export { createIdentitySlice, type IdentityAction } from './action';
