--- src/store/userMemory/slices/memory/index.ts
@@ -1,1 +1,0 @@
-export * from './action';
