+++ src/store/userMemory/slices/preference/index.ts
@@ -1,0 +1,1 @@
+export { createPreferenceSlice, type PreferenceAction } from './action';
